+++ src/translator/srcdiff_change.cpp
@@ -54,8 +54,8 @@ void srcdiff_change::output_whitespace_prefix() {
 */
 void srcdiff_change::output() {
 
-  unsigned int begin_original = rbuf_original->last_output;
-  unsigned int begin_modified = rbuf_modified->last_output;
+  int begin_original = rbuf_original->last_output;
+  int begin_modified = rbuf_modified->last_output;
 
   bool is_replace = false;
   if(end_original > begin_original && end_modified > begin_modified) {
@@ -76,7 +76,7 @@ void srcdiff_change::output() {
   if(end_original > begin_original) {
 
     bool first = true;
-    for(unsigned int i = begin_original; i < end_original; ++i) {
+    for(unsigned int i = begin_original; i < (unsigned int)end_original; ++i) {
 
       // output diff tag begin
       if(first && is_replace) {
@@ -145,7 +145,7 @@ void srcdiff_change::output() {
   if(end_modified > begin_modified) {
 
     bool first = true;
-    for(unsigned int i = begin_modified; i < end_modified; ++i) {
+    for(unsigned int i = begin_modified; i < (unsigned int)end_modified; ++i) {
 
       // output diff tag
       if(first && is_replace) {
